+++ gmp/solana/shared/Cargo.toml
@@ -4,3 +4,4 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+anchor-lang = "0.31.0"
